+++ programs/store/src/lib.rs
@@ -4337,3 +4337,6 @@ pub enum CoreError {
     #[msg("index overflow")]
     IndexOverflow,
 }
+
+#[cfg(not(feature = "no-entrypoint"))]
+gmsol_utils::security_txt!("GMX-Solana Store Program");